graphql-alt: run the framework's validations before our own
## Description

This ensures that basic validity checks (like fragments exist, and
variables are bound etc) are checked before our validation rules are
run, so the errors the service produces about these issues are specific
to the issue and not something our validation rule is trying to do that
relies on a correctness property.

This should improve the clarity of error messages for basic issues.

## Test plan

Existing tests, and run the service, trying the following query:

```graphql
{ ...IDontExist }
```